% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.drg.R
\name{statistics.drg}
\alias{statistics.drg}
\title{Statistics Calculation}
\usage{
statistics.drg(df, option = "mean")
}
\arguments{
\item{df}{a dataframe}

\item{option}{a string name for how to calculate (mean, median, sd)}
}
\value{
a dataframe of statistics calculation by option
}
\description{
This function is to calculate statistics over all of the DRG codes
for average Medicare payments
}
\examples{
drg <- read.scv('data/DRG_data.csv')
statistics.drg(drg, 'sd')

}
